pyproject.toml
[tool.poetry]
name = "react-langchain"
version = "0.1.0"
description = ""
authors = ["Darlington02 <darlingtonnnam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"
black = "^24.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
